// src/app/models/recibo-pagado.ts
export class ReciboPagado {
  cobrado: number;
  f_fact: string;
  fecha_pago: string;
  forma_pago: string;
  hora_pago: string;
  mes: Date;
  nis_rad: number;
  nom_agencia: string;
  recibo: string;
  sec_nis: number;
  sec_rec: number;
  tipo_recibo: string;
  vencimiento: string;
  est_rec: string;
}